requirements.txt
torch>=1.12.0
torchvision>=0.13.0
accelerate>=0.15.0
diffusers<=0.26.1
safetensors
einops>=0.6.0
hydra-core
loguru
matplotlib
numpy
omegaconf
opencv_python
Pillow
pyarrow
PyYAML
scikit_learn
tqdm
transformers>=4.25.1
pytorch_lightning
easydict